pyproject.toml
[project]
name = "dataengineering-journey"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "factory-boy>=3.3.3",
    "faker>=37.8.0",
    "psycopg2>=2.9.10",
    "pydantic>=2.11.10",
    "sqlalchemy>=2.0.43",
]
